Migrate cabal2nix's Main module to a library-exposed module.
The motivation is to support calling cabal2nix from another Haskell
library using the same interface provided by the executable.
