+++ cabal2nix.cabal
@@ -105,6 +105,7 @@ library
     , transformers
     , yaml
   exposed-modules:
+      Cabal2nix
       Distribution.Nixpkgs.Fetch
       Distribution.Nixpkgs.Haskell
       Distribution.Nixpkgs.Haskell.BuildInfo
